add googleMap ApiService

+++ app/src/main/java/fr/test200/findme/Place.kt
@@ -4,11 +4,11 @@ enum class RadiusType(val value: Int) {
     Small(0), Normal(1), Large(2), Big(3)
 }
 
-data class Place (
+data class Place(
     val id: Int = 1,
     val name: String,
-    val latitude: Float,
-    var longitude: Float,
+    val latitude: Double,
+    var longitude: Double,
     var difficulty: Int,
     val radius_type: RadiusType,
     val active: Boolean,
